Upgrade example to Expo 44
