+++ example/package.json
@@ -12,22 +12,27 @@
   },
   "dependencies": {
     "@expo/samples": "^3.0.3",
-    "@expo/vector-icons": "^10.0.6",
-    "expo": "^36.0.0",
-    "expo-font": "^8.0.0",
-    "react": "16.9.0",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-36.0.1.tar.gz",
-    "react-native-gesture-handler": "^1.5.6",
+    "@expo/vector-icons": "^12.0.0",
+    "@react-navigation/bottom-tabs": "^6.1.8",
+    "@react-navigation/native": "^6.0.6",
+    "@react-navigation/stack": "^6.0.11",
+    "expo": "44.0.0",
+    "expo-font": "10.0.4",
+    "react": "17.0.1",
+    "react-native": "0.64.3",
+    "react-native-gesture-handler": "2.1.0",
     "react-native-offline": "5.4.0",
-    "react-native-screens": "^2.0.0-beta.2",
-    "react-navigation": "^3.0.9",
+    "react-native-safe-area-context": "3.3.2",
+    "react-native-screens": "3.10.1",
     "redux": "^4.0.5",
     "redux-saga": "^1.1.3"
   },
   "devDependencies": {
+    "@types/react-native": "^0.60.15",
     "@types/react-navigation": "3.4.0",
     "babel-preset-expo": "^8.0.0",
-    "jest-expo": "^36.0.0"
+    "jest-expo": "^44.0.0",
+    "typescript": "^3.7.2"
   },
   "private": true
 }
